server/requirements.txt
django==3.0.7
pillow
djangorestframework
django-cors-headers
